test(config): restructure DTNI unittests into per-package suites
Split the four monolithic test_dtni_*.py files (flat in cvs/lib/unittests)
into per-source-module suites under per-package unittests/ dirs, matching the
convention already used by cvs/core and cvs/lib/{parallel,preflight}:

  cvs/lib/config/unittests/        test_base/loader/sweep/thresholds/migrate
  cvs/lib/config/frameworks/unittests/  test_vllm
  cvs/lib/runtime/unittests/       test_container_handle
  cvs/lib/unittests/               test_failure_taxonomy

Fixtures become framework-keyed (_fixtures.py: BASES, make_base, iter_bases)
so generic spine invariants (hashing, dispatch, env resolution, threshold-union
parsing) are parametrized over every framework base and re-prove automatically
as frameworks are added. Framework-specific rules (TP-vs-topology, vLLM sweep
axes, workload_kind) move to test_vllm.py. The sweep engine is exercised on
framework-neutral raw dicts.

No assertion dropped; adds vLLM concurrency/sequence_combinations validator
coverage. 63 tests pass; ruff format/lint clean.
